workaround for findpeaks
findpeaks.mat is a built-in matlab function and a function in the
chronux toolbox.for the waveform analysis, I used the buiilt-in function
initially but the chronux version is called when it is added to the
path. I added a workaround that removes the path to the chronux foilder
for the time of the function call.